feat(dashboard): integrate Bans page UI (design components)
